# app/Http/Controllers/FinancialReleasesController.php
<?php

namespace App\Http\Controllers;

use Illuminate\Http\Request;

use App\Service\FinancialReleasesService;
use App\Http\Requests\FinancialReleases\StoreRequest;
use OpenApi\Attributes as OA;


class FinancialReleasesController extends Controller
{

    public function __construct(
        private FinancialReleasesService $financialReleasesService
    ){}

    /**
     * Display a listing of the resource.
     */
    public function index()
    {
        //
    }

    /**
     * Show the form for creating a new resource.
     */
    public function create()
    {
        //
    }

    /**
     * Store a newly created resource in storage.
     */
    
    public function store(StoreRequest $request)
    {
    
        $response = $this->financialReleasesService->createFinancialRelease($request->validated());
    
        if($response) {
            return response()->json([
                'message' => 'Liberação financeira criada com sucesso',
                'data' => $response
            ], 201);
        } else {
            return response()->json([
                'message' => 'Erro ao criar liberação financeira'
            ], 500);
        }

    }

    /**
     * Display the specified resource.
     */
    public function show(string $id)
    {
        //
    }

    /**
     * Show the form for editing the specified resource.
     */
    public function edit(string $id)
    {
        //
    }

    /**
     * Update the specified resource in storage.
     */
    public function update(Request $request, string $id)
    {
        //
    }

    /**
     * Remove the specified resource from storage.
     */
    public function destroy(string $id)
    {
        //
    }
}